+++ components/src/core/Chip/Chip.tsx
@@ -10,6 +10,7 @@ import { Icon } from '../Icon';
 import { IconSource } from '../__types__';
 import { ExtendedTheme, useExtendedTheme } from '@brightlayer-ui/react-native-themes';
 import { fontStyleRegular } from '../Utility/shared';
+import { StyleProp, ViewStyle } from 'react-native';
 
 /**
  * Props for the Chip component.
@@ -68,6 +69,12 @@ export type ChipProps = Omit<PaperChipProps, 'icon' | 'mode' | 'selectedColor'> 
      * @prop {React.ReactElement} [avatar] - Avatar component to be displayed in the Chip.
      */
     avatar?: React.ReactElement; // New prop for passing Avatar component
+    /**
+     * Style overrides for internal elements. The styles you provide will be combined with the default styles.
+     */
+    styles?: {
+        root?: StyleProp<ViewStyle>;
+    };
 };
 
 export const Chip: React.FC<ChipProps> = (props) => {
@@ -85,6 +92,7 @@ export const Chip: React.FC<ChipProps> = (props) => {
         borderColor,
         textColor,
         theme: themeOverride,
+        styles = {},
         ...rest
     } = props;
 
@@ -153,6 +161,7 @@ export const Chip: React.FC<ChipProps> = (props) => {
 
     const renderCloseIcon = (): JSX.Element => <Icon source={{ name: 'close' }} size={18} color={DefaultTextColor} />;
     const renderIcon = (): JSX.Element | undefined => getIcon();
+
     return (
         <>
             {icon ? (
@@ -161,6 +170,7 @@ export const Chip: React.FC<ChipProps> = (props) => {
                         {
                             backgroundColor: chipColor ? chipColor : defaultChipColor,
                         },
+                        styles.root,
                         style,
                         chipStyle,
                     ]}
@@ -180,7 +190,9 @@ export const Chip: React.FC<ChipProps> = (props) => {
                     style={[
                         {
                             backgroundColor: chipColor ? chipColor : defaultChipColor,
+                            paddingVertical: avatar.props.size ? (avatar.props.size > 24 ? 4 : 0) : 0,
                         },
+                        styles.root,
                         style,
                         chipStyle,
                     ]}
@@ -188,7 +200,16 @@ export const Chip: React.FC<ChipProps> = (props) => {
                     showSelectedCheck={false}
                     selected={selected}
                     disabled={disabled}
-                    avatar={avatar}
+                    avatar={React.cloneElement(avatar, {
+                        style: avatar.props.size
+                            ? {
+                                  height: avatar.props.size,
+                                  width: avatar.props.size,
+                                  borderRadius: avatar.props.size,
+                                  ...avatar.props.style,
+                              }
+                            : {},
+                    })}
                     {...(isElevated && { elevated: !disabled })}
                     closeIcon={renderCloseIcon}
                     {...rest}
@@ -201,6 +222,7 @@ export const Chip: React.FC<ChipProps> = (props) => {
                         {
                             backgroundColor: chipColor ? chipColor : defaultChipColor,
                         },
+                        styles.root,
                         style,
                         chipStyle,
                     ]}