+++ components/src/core/Drawer/Drawer.tsx
@@ -50,17 +50,19 @@ export const Drawer: React.FC<DrawerProps> = (props) => {
     const {
         // Inheritable Props
         /* eslint-disable @typescript-eslint/no-unused-vars */
+        activeChevronColor,
         activeItemBackgroundColor,
-        activeItemBackgroundShape,
+        activeItemBackgroundShape = 'square',
         activeItemFontColor,
         activeItemIconColor,
         backgroundColor,
-        chevron,
+        chevron = false,
+        chevronColor,
         collapseIcon,
         disableActiveItemParentStyles,
-        divider,
+        divider = false,
         expandIcon,
-        hidePadding,
+        hidePadding = true,
         itemFontColor,
         itemIconColor,
         nestedBackgroundColor,
@@ -76,6 +78,14 @@ export const Drawer: React.FC<DrawerProps> = (props) => {
         ...viewProps
     } = props;
 
+    const defaultProps: Partial<DrawerProps> = {
+        activeItemBackgroundShape: 'square',
+        chevron: false,
+        divider: false,
+        hidePadding: true,
+        styles: {},
+    };
+
     const theme = useExtendedTheme(themeOverride);
     const insets = useSafeAreaInsets();
 
@@ -88,7 +98,7 @@ export const Drawer: React.FC<DrawerProps> = (props) => {
                 .map((child) => {
                     let inheritableProps = {};
                     if (inherit) {
-                        inheritableProps = inheritSharedProps({ ...props, theme }, child.props);
+                        inheritableProps = inheritSharedProps({ ...defaultProps, ...props, theme }, child.props);
                     }
                     return React.cloneElement(child, inheritableProps);
                 }),
@@ -110,11 +120,4 @@ export const Drawer: React.FC<DrawerProps> = (props) => {
             </Surface>
         </DrawerContext.Provider>
     );
-};
-
-Drawer.defaultProps = {
-    activeItemBackgroundShape: 'square',
-    chevron: false,
-    divider: false,
-    hidePadding: true,
 };